Massive restructure for Account service

+++ services/authentication/server/src/main/service/authentication-service.js
@@ -215,6 +215,9 @@ class AuthenticationService {
   }
 
   async #logoutViaRefreshToken(authority, refreshToken) {
+    if (typeof refreshToken !== 'string') {
+      return;
+    }
     try {
       await this.#persistentSessionRepository.deleteByRefreshToken(refreshToken);
     }
